+++ angr/analyses/cfg/cfg_utils.py
@@ -1,0 +1,1 @@
+from angr.utils.graph import GraphUtils as CFGUtils
