+++ angr/analyses/decompiler/decompiler.py
@@ -21,7 +21,6 @@ from .decompilation_options import DecompilationOption
 from .decompilation_cache import DecompilationCache
 from .utils import remove_labels
 from .sequence_walker import SequenceWalker
-from .ailblock_walker import AILBlockWalkerBase
 
 if TYPE_CHECKING:
     from angr.knowledge_plugins.cfg.cfg_model import CFGModel
@@ -447,7 +446,7 @@ class Decompiler(Analysis):
             const_values.add(expr.value)
 
         def _handle_block(block: ailment.Block, **kwargs):  # pylint:disable=unused-argument
-            block_walker = AILBlockWalkerBase(
+            block_walker = ailment.AILBlockWalkerBase(
                 expr_handlers={
                     ailment.Expr.Const: _handle_Const,
                 }
