+++ angr/analyses/decompiler/optimization_passes/const_derefs.py
@@ -2,12 +2,11 @@
 from typing import Optional, TYPE_CHECKING
 import logging
 
-from ailment import Block
+from ailment import Block, AILBlockWalker
 from ailment.expression import Load, Const, BinaryOp, UnaryOp
 from ailment.statement import Statement, Assignment, Call, ConditionalJump
 
 from ..ailgraph_walker import AILGraphWalker
-from ..ailblock_walker import AILBlockWalker
 from .optimization_pass import OptimizationPass, OptimizationPassStage
 
 if TYPE_CHECKING:
