+++ angr/analyses/proximity_graph.py
@@ -4,12 +4,12 @@ import logging
 import networkx
 
 import ailment
+from ailment import AILBlockWalker
 
 from ..codenode import BlockNode
 from ..sim_variable import SimMemoryVariable
 from ..knowledge_plugins.functions import Function
 from .analysis import Analysis, AnalysesHub
-from .decompiler.ailblock_walker import AILBlockWalker
 
 if TYPE_CHECKING:
     from angr.analyses.decompiler.decompiler import Decompiler
